+++ hal/gles/adapter.go
@@ -44,11 +44,19 @@ func (a *Adapter) Open(_ gputypes.Features, _ gputypes.Limits) (hal.OpenDevice, 
 	vao := a.glCtx.GenVertexArrays(1)
 	a.glCtx.BindVertexArray(vao)
 
+	// Query hardware texture unit limit for binding validation.
+	var maxTexUnits int32
+	a.glCtx.GetIntegerv(gl.MAX_TEXTURE_IMAGE_UNITS, &maxTexUnits)
+	if maxTexUnits <= 0 {
+		maxTexUnits = 8 // Conservative default
+	}
+
 	device := &Device{
-		glCtx:  a.glCtx,
-		wglCtx: a.wglCtx,
-		hwnd:   a.hwnd,
-		vao:    vao,
+		glCtx:           a.glCtx,
+		wglCtx:          a.wglCtx,
+		hwnd:            a.hwnd,
+		vao:             vao,
+		maxTextureUnits: maxTexUnits,
 	}
 
 	queue := &Queue{